reorganize file preprocess and preprocessor_utils
